Remove force unwrap rule (we know what we're doing :) )
Our use of force unwrapping is similar to index-out-of-bounds. For
example, Reflection makes heavy use of force unwrapping when creating
pointers, etc. If it was to silently fail or throw an error instead,
it would simply mean that the app would not work, which is equivalent
to (if not worse than) the app crashing.